+++ CHANGELOG.md
@@ -7,11 +7,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Configuration for first day of the week #950
 - StyleCI configuration #929
+- Added expense category
 
 ### Changed
 - Removed `invoiceninja.komodoproject` from Git #932
 - `APP_CIPHER` changed from `rinjdael-128` to `AES-256-CBC` #898
-- Improved options when exporting data 
+- Improved options when exporting data
 
 ### Fixed
 - "Manual entry" untranslatable #562